Add complete IR dashboard
