+++ register/api/registration/cr.py
@@ -1,46 +1,15 @@
+from register.api.registration.types.registration import get_registration
 from util import JSONError, get_user, status
-
-from rest_framework.parsers import JSONParser
 
 from django.http import JsonResponse
 from django.views.decorators.http import require_POST
 
-from register.api.registration.types.registration import Registration
-from register.api.registration.types.util import get_serializer
+from register.api.registration.types.util import get_serializer, put_registration
 
 from register.models import SignupLog
 from accounting.models import Order
 
 from register.views import send_CR_confirmation_email
-
-
-def put_cr_registration(request, registration, purchasing_company):
-    data = JSONParser().parse(request)
-    serializer = get_serializer(
-        registration,
-        context={"user": get_user(request), "purchasing_company": purchasing_company},
-        data=data,
-    )
-
-    if serializer.is_valid():
-        serializer.update(registration, serializer.validated_data)
-        return JsonResponse(serializer.data)
-
-    return status.serializer_error(serializer.errors)
-
-
-def get_registration(company, fair, contact, exhibitor):
-    orders = Order.objects.filter(
-        purchasing_company=company, product__revenue__fair=fair
-    )
-
-    return Registration(
-        company=company,
-        contact=contact,
-        fair=fair,
-        exhibitor=exhibitor,
-        orders=orders,
-    )
 
 
 def handle_cr(request, company, fair, contact, exhibitor):
@@ -54,7 +23,7 @@ def handle_cr(request, company, fair, contact, exhibitor):
     if request.method == "GET":
         return JsonResponse(serializer.data, safe=False)
     elif request.method == "PUT":
-        return put_cr_registration(request, registration, company)
+        return put_registration(request, registration, company)
     else:
         return status.UNSUPPORTED_METHOD
 
@@ -71,13 +40,13 @@ def submit_cr(request, company, fair, contact, exhibitor):
     if registration.contact == None:
         return status.COMPANY_DOES_NOT_EXIST
 
-    signature = registration.signature
+    signature = registration.cr_signature
 
     if signature != None:
         return status.EXHIBITOR_ALREADY_SIGNED
 
     signature = SignupLog.objects.create(
-        company_contact=contact, contract=registration.contract, company=company
+        company_contact=contact, contract=registration.cr_contract, company=company
     )
 
     # Add package products
