+++ register/api/registration/types/cr.py
@@ -3,7 +3,8 @@ from accounting.models import Order
 from accounting.api import OrderSerializer
 
 from companies.serializers import CompanySerializer
-from register.api.registration.types import RegistrationSerializer
+from register.api.registration.types.serializer import RegistrationSerializer
+
 from util import update_field
