+++ register/api_urls.py
@@ -6,6 +6,7 @@ app_name = "register_api"
 
 urlpatterns = [
     url(r"^$", registration.index),
-    url(r"^submit$", registration.submit),
+    url(r"^sign_ir$", registration.sign_ir),
+    url(r"^submit$", registration.sign_cr),
     url(r"^(?P<company_pk>[0-9]+)$", registration.get_company),
 ]
